Improved security of ssh-tunneld
Added an option to control binding of the "control socket". Default is
to bind only to local interface (127.0.0.1). Specifying -r on the
command line binds to * and allows remote control connections. Note that
the SSH proxy port is not affected by this. For that, you need to change
the Gateway options in your ssh configuration.