Improved code readability
Added initialising values to most variables. Also altered the order of
if() .. elseif () .. else in the code to fork and become session leader,
so the simple cases are dealt with first (parent, error --- both of
which cause an _exit) then the child code can be outside the if block.
This improves locality of the logic when parsed by a human reader.
Previously the code to deal with the trivial parent or error cases came
after the somewhat longer child code, and it was easy to forget the
initial condition!