Add checks for NULL after each malloc() and strdup()
There is only one call to malloc, so check for NULL there.
Several calls to strdup() exist in the options.c files for
ssh-tunneld and ssh-tunnelc, so wrap strdup() in a function
called checked_strdup(), and call that everywhere else.

checked_strdup() is defined once per project (ssh-tunneld
and ssh-tunnelc) so as to keep the two codebases separated.