Ensure correct null-termination of strings made by strncpy()
One location does this correctly anyway, so added a comment to that
effect. The code in ssh-tunnelc/logging.c could, in theory, fail to
terminate the string properly, and a subsequent strlen() would
cause problems. This is fixed by defining a new function,
terminated_strlen() which always terminates, even if it means
chopping a bit off the end of the string!